Make GenParticleProducer a global module
GenParticleProducer was showing up as stalling during GEN+SIM
multi-threaded jobs. Removing the use of member data as temporary
storage during the call to produce allowed the module to be a
global module. The fast lookup of PDGID to charge was moved to a
run cache. Given that the vast majority of GEN+SIM jobs only
process one run, doing the calculate once per run instead of once
per job should not be a problem.