Made BaseEvtVtxGenerator a stream module
In addition to making BaseEvtVtxGenerator a stream module, the
virtual functions of the class were changed to be const. This was
originally done in an attempt to make the module a global module.
That was abandoned since some base classes were using beginRun to
read from the EventSetup and set their parameters (which could be
done using RunCache but would complicate the design). The member
functions were kept const since I believe they allowed the implementation
to be better.