Comply to CMSSW coding rules.
The try-catch clause is redundant here because the only exception
which could be thrown is if the memory allocation for the FileBlob
content does not succeed, i.e. we are out of RAM.
If this happens, the job should quit anyways, I think.

Every other failure is covered by the check for the FileBlob size.