Move sortByStringFunction to StringObjectFunction.h.
PhysicsTools/UtilAlgos/interface/CachingVariable.h is sometimes only
included because of the sortByStringFunction template. This include
however is causing a layering violation because for example
CommonTools/UtilAlgos and PhysicsTools/UtilAlgos now include each
other.

This patch moves the sortByStringFunction template into the
StringObjectFunction.h and removes the now unnecessary includes
to CachingVariable.h and in StringBasedNTupler.h.