Don't forward declare CaloTower in CaloTowerFwd.h
This header won't compile with just a forward declaration of
CaloTower, because the Ref and RefVector types we declare will check the
ValueTrait of it's containing class, which is in this case CaloTower.
ValueTrait needs to have the definition of CaloTower to work, but
we only have a forward declaration in this header available.

This leads to undefined behavior, so we should replace this forward
declaration with an actual include.

People seem to anyway always include CaloTower too to
avoid this cryptic error, so I don't think this will hit compilation
times that much.