Fix includes/fwd decls in Fireworks/Core
We have to forward declare TEveWindowSlot/TEveElement because
we reference pointers to those types in the respective headers.

We have to include Rtypes.h to get Color_t in FWDetailViewBase.h
and FWModelExpressionSelector.h. In FWViewBase.h we have to
include Rtypes.h for Int_t.

In FWDetailViewManager.h/FWInteractionList.h we have to
include <vector> because we use std::vector there.