bug fix: make sure product registry is updated before reading
The previous code was calling the product registry directly
and assumed calling EventBase::triggerNames would always update
the registry. That assumption was incorrect. The registry was
only updated if the triggerNames PSet was not found in the
cache. Calling EventBase::parameterSet does guarantee that
the registry is up to date.