Calibration/EcalCalibAlgos: Fix clang warning [-Wtautological-constant-out-of-range-compare]
warning: comparison of constant 140 with expression of type 'reco::GsfElectron::Classification' is always true [-Wtautological-constant-out-of-range-compare]
zeeCandidates[myBestZ].second->getRecoElectron()->classification()!= 140);
~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~^  ~~~

Store return value on ClassificationVariables::classification() in a temporary variable and then compare to the "magic numbers". This is obviously an unused algorithm because the conversion to the classification enum values was never made. I think this algorithm was implemented elsewhere in the code. This aviods the clang warning, but if this is unused code it should be removed.