Use electronGsfTracksFromMC
For ecalDrivenGsfElectronCoresFromMC we need to use the gsfTracks
track collection in order to make Ref for electrons in the final
electrons collection, since this is the only collection that has been
directly produced consuming the HGCal Multi Clusters. Failing to do
so, we would end up using the PFTrack collection that, instead, is
using the (realistic) simCluster information.