Fix circular dependency in validation_step path
The sequence bdHadronTrackValidationSeq was causing a circular module
dependency in the validation_step path. This was not noticed originally
because the modules causing the circular dependency were being run
in another path. Changing bdHadronTrackValidationSeq to use a Task
solved the problem. This did not make a substantial change to the
runtime order of modules because mucn of that sequence's modules
were already on other Tasks.