Fix circular dependency in pathALCARECOEcalUncal*Electron paths
An improvement to the module circular dependency finder uncovered
a circular dependency in all pathALCARECOEcalUncal*Electron. The
Fix was to move the problematic modules to a Task.

The underlying problem was
-ecalDigis was after WenuSelector on the ALCA paths
-WenuSelector consumes pfMet
-pfMet is after particleFlowRecHitECAL on path reconstruction_step
-particleFlowRecHitECAL consumes ecalDigis
This only happened to work because ecalDigis is also on the path
raw2digi_step. However, if there was an exception on raw2digi_step,
ecalDigis would not be run from that path and then the job would
deadlock because of the remaining circular dependency.