Bug fix. Add 2 overrides to LHESource
Add readRunAuxiliary_ and readLuminosityBlockAuxiliary_
function overrides to LHESource. (also includes code-check
fixes and indentation fixes that remove inconsistencies and
tabs, but the only substantive changes are the 2 new functions).
Earlier changes in the Framework broke LHESource and the
addition of these two functions fixes the problem by setting
the ProcessHistoryID properly.