Have TTreeCache work with EmbeddedRootSource
The EmbeddedRootSource calls different functions of RootFile and
RootTree. This was causing it to not properly use the TTreeCache.
Now we will use the TTreeCache in that case and properly handle
the case where we start reading near the end of the file and then
jump back to the beginning before training has completed.