Recover CaloParticle into SimPFProducer
Every CaloParticle is linked to the first stable particle originating
the cascade of particles that left hits in the calorimeters: that
stable particle is not included as a simCluster (unless it itself left
hits in the calorimiters). Additional code has been added to find a
matching track based on the trackId linked to the CaloParticle itself,
not to the associated simClusters. The most important clients affected
by this change are electrons.