Test ROOT Schema Evolution on TH*
When an update on the ROOT external brings in schema evolution for any
object that has to be handled also by DQMGUI, trigger a test failure.
The fix of the test is trivial. This will serve as a check-and-trigger
to update the root.spec file on the comp side of cmsdist.