Add GlobalCache to DeepFlavourJetTagsProducer to improve startup performance
igprof profiles showed this module was taking significant
time at startup, most as it reads the configuration file
used to initialize its neural network (lots of memory
churn also). Without the GlobalCache the startup time
is proportional to the number of streams. This change
moves the LightweightNeuralNetwork into the global cache
so there is only one and the initialization time is
no longer proportional to the number of streams.
LightweightNeuralNetwork is supposed to be thread safe
to use in this manner.