Explicitly request "SoftDropPuppi" subjets
This addresses the issue raised in https://github.com/cms-nanoAOD/cmssw/issues/194. In the past, two subjet collections were present ("SoftDrop" and "SoftDropPuppi"). We eventually dropped the first, and the code was written after that point so I didn't think about it. However, running on 2016 data or MC will access the first collection. These indices can be out of sync if a subjet was dropped by PUPPI but not by CHS. It is rare, so not caught in unit tests.

This can be worked around by requiring that the delta R between the subjet and the AK8 jet be less than 0.8 to remove these corner cases. If you want an alternative, just do a delta R matching between the subjets and AK8 jet to begin with, which is also safe.