New trace tool based on the configuration content.
This is related to the PSetHistory mode in CMSSW, but tracks something slightly different: The goal is to keep track of all places in the configuration that actually affect the process running in the end.

A the moment there is no output, but the _trace_events on each Sequence are populated with the lines of code that modified this Sequence. Since all the constituents are preserved in the _seq itself, this should be enough to reconstruct the entire configuration tree and assign a source location to each item.