Make separate plots browsable together
If Validation plots are produced using the 'separate' option, a static
web page will be created with the proper links to all the images. This
commit will add in that specific directory the index.php file, and all
its needed components, to be able to browse the content of the
directory automatically, hence visualizing all images at a glance.