Improve exception handling for end stream exceptions
These improvements are particularly relevant when
running with concurrent lumis.

More thread safe. Just use an atomic<bool> for exceptionMessageLumis_.
No reason to pass around the string. It's always the same.

Set the deferredExceptionPtr_ when there are exceptions in end lumi
transitions to shut down quicker and with less noise in the printed
exceptions.

Some cleanup to slightly simplify this part of the code.