Coerce the FwkIO unit tests into passing again...
... by emulating the per-lumi reset in in the DQMStore.
Ideally we'd test for the new behaviour of harvesting instead of
emulating the old behaviour for the tests, but that would break
the assumption in the tests that saving and loading DQMIO files
does not change the contents.
We could keep this (and solve many other problems) by keeping
the per-lumi histos and their per-run harvested versions
separate in the DQMStore, but the legacy-mode DQMStore in
Harvesting really does not understand such concepts. So, we have
to admit defeat and cheat in the tests until we can run
harvesting with a properly structured DQMStore.

The forceResetOnBeginLumi option was not used anywhere, so
changing the behaviour to something more sane should be safe;
expanding the per-run reset to per-lumi histos makes sense
in general as well.