ESGetToken now uses indices
The ESGetToken now just contains indices, not the strings from an ESInputTag. These indices are used to lookup the index for DataProxies in the appropriate EventSetup Record. The proxy index is then used to directly call the Proxy.
The lookup tables from token index to proxy index are filled at begin job time.