Allow parameters in a PSet to be changed via a dict
A dict can now be passed to PSet.setValue (which is also called during assignment). The key, value pairs in the dict are then iterated over and used to call assignment on the corresponding parameter labels within the PSet. If the PSet contains other parameters than those with keys in the dictionary, those other parameters are not changed.