Avoid unrepresentable int value in BitSet
The bit manipulation used to iterate over the BitSet required
a negative value to be negated and the number obtained could not
be represented by an int when the bit in question was # 31.
The new bit manipulation gives the same results (as tested in the
unit test) but avoids the undefined behavior.

This was found by UBSAN.