Make getTH1() and friends non-const...
... with a workaround for dqm::legacy, where it is still accessible as const.
This was needed (among others) to keep all the QTest logic compileable.

Lots of messing around with namespaces to get everything to line up; we use
the legacy ME for all but DQMGlobalEDAnalyzer now.

Compiles quite a lot of CMSSW, but not all.
This legacy interface is probably not save to use, but then it is to keep supporting things that where unsafe before.