Remove DumpFile, MergeFile and DQMCollector.
While probably useful, I don't think these tools are used much today.
We cannot support these heavily 'DQMstore as a Service' based tools in the future, so let's remove them already and see.
DQMCollector is probably actively used in the GUI, but this is only the launcher, not the actual DQMNet implementation.
Also the GUI uses a fixed version of these files from CMSSW7_6_0, it is effectively diverged.