Make SiStripOfflineDQM explicitly depend on its QualityTester.
This should make sure that the QReports are available by the time that
SiStripOfflineDQM analyzes them in endRun.

Except, it seems to not work, and the modules still run in reverse
order.

The change to edm::one is to take the more complicated edm legacy
semantics out of the equations, but does not seem to change anything.

Includes the instance labels on suggestion of @Dr15Jones.