Use stable sort when reading DQMIO.
This is mostly to make the tests happy: When reading pre-run scalars from split DQMIO files, the result could be the value from any file, since it dependes on processing order.
We can't do anything about that in general (we might have a job processing lumi 1 and 4, and another 2 and 3, and then merge the results; which value should be used?),
but for the tests we can at least make the behaviour detetministic for the test case.