Add header guards for DYTThrObject.h and Serializable.h
For Serializable.h, we should change #pragma once, because it gets the filename and creates a include guard based on it.
If it relocates it becomes problematic, because it is could be possible to have two files with the same name in two different locations.
(to avoid such cases we should rely less on #pragma once)