Added links to inspect bin by bin comparison with a raw rootjs viewer
Sometimes there are changes to DQM files that are not fully visualised
by the DQM GUI. For example when efficiency flag was added/removed or
when a QT test result changes but the ME that was tested does not. This
change allows to inspect resulting comparison files without treating them as DQM files.