Add fraction of associated energy
For every CaloParticle that has been associated to a
Trackster/MultiCluster, add a plot showing the fraction of the shared
energy. The plot is filled once per CaloParticle by taking the
Trackster/MultiCluster with the best score. This plot is useful to judge
the quality of the score threshold in terms of effective shared energy.