Get MagneticField in beginRun()
Given that the ESGetToken is declared such that MagneticField is consumed in beginRun, the product must be read from EventSetup always there. Alternative would be to have two tokens (one for beginRun and another for produce), but always reading in beginRun is simpler and along the planned evolution of this code.