Improve HGCAL geometry handling in Fireworks
Make use of the topology information, in the form of 9 floats for each
DetId, to pass HGCAL specific information to Fireworks. This will avoid
the weak and error prone practice of unpacking the DetId bits while
rendering the objects via the dedicated proxies. Three floats are still
available. This will increase the coupling between the release, the
geometry dumping and the explored file. On the other hand, at present,
the unpacking is wrong for new geometries.