Speed up interactive dictionary generation for ElectronMVAEstimatorRun2
Now that the dictionaries for the ElectronIdentification package is
removed, there is still one usecase for which we need to find an
alternative solution. The `ElectronMVAEstimatorRun2` class needs to be
usable from Python in `ElectronIdentification/python/FWLite.py`,
which we can achieved by including `ElectronMVAEstimatorRun2.h` with the
ROOT interpreter in Python. However, doing so was very slow because of
all the includes in `ElectronMVAEstimatorRun2.h` and
`AnyMVAEstimatorRun2Base.h`, and there were even errors caused by
conflicting modules on my system.

Hence, most include statements in `ElectronMVAEstimatorRun2.h` were
replaced by forward declarations in this commit. Now the call to
`gInterpreter` returns almost instantaneously.