Improved PFID model selection consistency.
The model index used to evaluate the candidate is now saved in the DNNHelper output and used in the producer
to select how many DNN outputs should be saved, without performing again the pt/eta binning.

Moreover the eta selection is now performed with SuperCluster.eta instead of Electron.eta.