First mkFit runs with Phase2 data / geometry!
* mkFit.cc: Make use-dead-modules really optional. They are hardcoded
for CMS-2107, some unknown setup.

* MkBuilder.cc: Make some high-volume debug output optional.

* val_scripts/validation-cmssw-benchmarks-multiiter.sh:
- Add new sample TTbar_Phase2.
- Add variable for numiters
- Make --use-dead-modules configurable
- Add var for exstra mkFit args to pass in --geom CMS-2029
- Detect Fedora and use lower n_threads / maxvu. Use AVX2 instead
  of AVX_512. This should be configurable in some other way.

* plotting/Common.hh: Do not try to plot STD in MIMI case - exported
via an env variable for lack of a better idea.
We should support STD in MIMI with upcoming configurability changes.

* validation-desc.txt: Add a prelimiary note describing how I ran
the Phase2 test. Standalone validation has not been tried since the
move into CMSSW so this should be worked on further.

* xeon_scripts/init-env.sh:
Detect Fedora and use el8 binaries in this case.
This should be the default once we drop Centos-7.