Autodetect the Pixel FED size
Allocate a WordFedAppender for each event using non-write-conbined,
pinned memory (generally from the host caching allocator) based on the
actual size of the FEDs.

This makes the MaxFEDWords configuration parametere obsolete. It is
still kept for backwards compatibility with older configurations, but it
will not be written to the cfi files and will be ignored if present.