Improve convertToRaw
After processing each run, write a cff.py file with the Source and the
necessary Services to process the converted files.

Improve the handling of the output and exit status of the subprocesses,
and add a -v/--verbose option to print additional message while
preprocessing and converting the input files.