Bug fix, rare race to insert into streamQueuesInserter_
There is the possibility for stream begin lumi tasks
to be inserted into the stream queues before the
stream begin run tasks. This is very rare and it takes
some high statistics to hit the bug. The problem is fixed
by moving one line of code later. As a side effect of that
a pause is needed to ensure stream begin run executes before
stream begin lumi in the single stream case with this fix.

Bug was just introduced with run concurrency, noticed in the
IBs, and immediately fixed.