Add a test for CUDA libraries and build rules
The package HeterogeneousTest/CUDADevice implements a library that
defines and exports CUDA device-only functions, and a plugin and test
that use them.

The package HeterogeneousTest/CUDAKernel implements a library that
imports device functions from HeterogeneousTest/CUDADevice to define
and export CUDA kernels, and a plugin and test that use them.

The package HeterogeneousTest/CUDAWrapper implements a library that
imports kernels from HeterogeneousTest/CUDAKernel to define and
export host-only wrappers around them, usable by non-CUDA libraries,
plugins and applications, and implements a plugin and test that use them.

The package HeterogeneousTest/CUDAOpaque implements a library that
imports kernels from HeterogeneousTest/CUDAKernel to define and
export host-only wrappers around the whole CUDA section, usable by
libraries, plugins and applications that are not CUDA-aware, and
implements a plugin and test that use them.