Clean BuildFiles
Another quick BuildFile cleaning PR in the style of many before
(for example https://github.com/cms-sw/cmssw/pull/42336).

In some cases, the dependencies had to be moved to the
appropriate BuildFile, for example from a plugins directory
to the library directory or vice versa.

This PR cleans unnecessary includes from CMSSW BuildFiles that
were added in the last months.