Fix random DQM path order from ConfigBuilder in case of duplicate DQM sequences
PromptReco for Heavy Ions was generating different RECO configurations
for each job of the same Primary Dataset (and Run). This caused
problems during merging as the ParameterSet registry grew.

The difference between the RECO configurations was in how the DQM
sequences were assigned to DQM Paths. One source for random shuffling
was found to be in ConfigBuilder in case the expansion of the DQM
sequence specifier leads to duplicate sequences.

This commit fixes the path ordering, and adds a test for it.

Co-authored-by: Christopher Jones <chrisdjones15@gmail.com>