Make empty std::optionals to work as ESProducer return value
The std:optional as an ESProducer::produce() return type has been
supported for a long time already. However, the ESProducer::produce()
returning an empty std::optional lead to an exception being thrown
(about de-referencing an empty std::optional). The fix is use the the
default form of the moveFromTo() function also for std::optional.

I removed the overload for std::unique_ptr as well since it provided
no functionality over the default implementation of the function.