Use likelihood fit instead of chi-square fit in DQMServices slice fits
The DQM plots use the `TH2::FitSlicesY()` function to fit some Gaussians.
However, some of the fits are failing. This was not resulting in errors
so far, but with the switch to Minuit2 by default in ROOT 6.30 it will.

The problem is that it uses chi-square fits to fit slices with many
empty bins, which is not appropriate. Doing a likelihood fit with the
`"l"` option is one way to fix the problem, because it can better deal
with empty bins.

Closes #42979.