Add Alpaka producer that converts HCAL/ECAL hits to CaloRecHitSoA format
This commit adds a module that converts collections of CPU calorimeter hits (HBHERecHit or EcalRecHit) to the portable CaloRecHitSoA format. This step is necessary to process these hits in an Alpaka implementation of the ParticleFlow RecHit producer. Just like the CaloRecHitSoA format itself, this conversion is a temporary solution that will become obsolete, as soon as Alpaka-compatible SoAs are provided directly by the HCAL/ECAL groups.
The commit includes a configuration file, that executes the producer. This will serve as a basis for validation that will be introduced in future commits.