Implement PortableObject types
PortableHostObject<T> is a class template that wraps a struct type T and an
alpaka host buffer, that owns the memory where the struct is allocated. The
content of the struct is persistent, while the buffer itself is transient.
Specialisations of this template can be persisted, but requre special ROOT read
rules to read the data into an alpaka memory buffer.

PortableDeviceObject<T, TDev> is a class template that wraps a struct type T and
an alpaka device buffer, that owns the memory where the struct is allocated.
To avoid confusion and ODR-violations, the PortableDeviceObject<T, TDev>
template cannot be used with the Host device type.
Specialisations of this template are transient and cannot be persisted.

Add TestHostObject and TestDeviceObject based on a TestStruct.

Co-authored-by: Adriano Di Florio <adriano.diflorio@ba.infn.it>
Co-authored-by: Breno Orzari <breno.orzari@hotmail.com>
Co-authored-by: Dimitris Papagiannis <nothingface0@gmail.com>