Fix ordering of modules in HLT_TriMu_10_5_5_DZ_FromL1TkMuon path.
The ordering of the modules as currently defined in this path is
severely buggy and will result in an unrunnable schedule, if this path
is the only one scheduled. Once again, this problem surfaced while using
the new L1T menu in 14_2_0_pre1, but the scheduling problem is there
since a while, but went unnoticed so far.