Pause MaxMemoryPreload for edmplugincache file reading in PluginManager
In a local developer area with many packages checked out, the
PluginManager cache can be tens of megabytes larger than in a
developer area without packages being checked out. This difference can
result in false positive warnings in MaxMemoryPreload monitoring in PR
tests. An easy way (even if hacky) out is to exclude the PluginManager
cache from the MaxMemoryPreload monitoring.