fix retry logic for triton client
On retry the client was trying to access the TritonService through
the ServiceRegistry. However, the thread calling the evalute method
did not have the appropriate context setup to allow this. We now
save the ServiceToken when the client is created, so the appropriate
context can be setup before accessing the service.