Remove EMH physics list
It was added in October 2022 by commit dedcaad80a to try out G4HepEm
via the G4HepEmProcess interface. For maximum gain and complete physics
(including gamma-lepto-nuclear interactions), it is better to integrate
via the G4VTrackingManager interface as implemented by commit ab896a306d
in March 2023 (and soon to be updated).