Add OneToOne Table for Associators.
After OneToMany, the association of type OneToOne has been added as
well. The logic is surprisingly similar to the ordinary case of a
SimpleFlatTableProducer but for the external loop in the vector and the
automatic creation of the iota for the source index. This latter point
is still open for discussion, i.e., I'm not 100% convinced it is needed.