Extend CMSSW to a distributed application over MPI
Let multiple CMSSW processes on the same or different machines coordinate event
processing and transfer data products over MPI.

The implementation is based on four CMSSW modules. Two are responsible for
setting up the communication channels and coordinate the event processing:
  - a "remote controller" called MPIController;
  - a "remote source" called MPISource;
and two are responsible for the transfer of data products:
  - a "sender" called MPISender;
  - a "receiver" called MPIReceiver.

Data products can be serialised and transferred using the trivial serialisation
from HeterogeneousCore/TrivialSerialisation - if available - or the ROOT-based
serialisation.

Various tests are used to validate the implementation: matching the local and
remote event id, transferring SoA products with trivial serialisation, and
transferring legacy product with ROOT serialisation.

Co-authored-by: Andrea Bocci <andrea.bocci@cern.ch>
Co-authored-by: Anna Polova <anna.polova@cern.ch>
Co-authored-by: Mario Gonzalez <mario.gonzalez.carpintero@cern.ch>